tests/wpt/web-platform-tests/html/canvas/element/path-objects/2d.path.lineTo.nonfinite.details.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.lineTo.nonfinite.details</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.lineTo.nonfinite.details</h1>
<p class="desc">lineTo() with Infinity/NaN for first arg still converts the second arg</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/clear-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("lineTo() with Infinity/NaN for first arg still converts the second arg");
_addTest(function(canvas, ctx) {

for (var arg1 of [Infinity, -Infinity, NaN]) {
  var converted = false;
  ctx.lineTo(arg1, { valueOf: function() { converted = true; return 0; } });
  _assert(converted, "converted");
}


});
</script>
